Merge data-structure changes with new `PackedInstruction`
This is a heavily conflicted merge, as large tracts of the data model
have changed from underneath this branch.  This changes all the uses of
old from-Python conversion methods to the new `OperationFromPython`, and
uses the new forms of construction for `PackedInstruction` and
`CircuitInstruction` (i.e., mostly just building them directly).

Uses of the old `OperationType` are converted into their equivalent
forms on `PackedOperation`/`OperationRef`

Going beyond a simple merge, this adds an implementation of `Operation`
to `PackedOperation` directly (rather than needing to go via
`OperationRef`) to reduce the number of conflicted lines in the
`DAGCircuit` implementation, and to fit the usage more naturally.