Add DAG typed node support to C API. (#15206)
* Typed node interface for the DAG C API.

Adds the following supporting functions for working with nodes:

  * `qk_dag_num_op_nodes` => returns the number of operation nodes in the DAG.
  * `qk_dag_clbit_out_node` => given a clbit, returns the node index of its wire's output.
  * `qk_dag_clbit_in_node` => given a clbit, returns the node index of its wire's input.
  * `qk_dag_qubit_out_node` => given a qubit, returns the node index of its wire's output.
  * `qk_dag_qubit_in_node` => given a qubit, returns the node index of its wire's input.
  * `qk_dag_wire_node_value` => given a wire endpoint node index, returns the wrapped value (e.g. the qubit).
  * `qk_dag_op_node_num_qubits` => given an op node index, return the op's number of qubits.
  * `qk_dag_op_node_num_clbits` => given an op node index, return the op's number of clbits.
  * `qk_dag_op_node_num_params` => given an op node index, return the op's number of params.
  * `qk_dag_op_node_qubits` => given an op node index, return a pointer to the op's qubits.
  * `qk_dag_op_node_clbits` => given an op node index, return a pointer to the op's clbits.
  * `qk_dag_op_node_gate_op` => given an op node index of a gate operation, returns the op's `QkGate` and params.
  * `qk_dag_op_node_kind` => given an op node index, return the operation's kind.

Also adds a new `qk_dag_apply_gate` function to set the pattern we should use
for other DAG apply op functions. The function takes a `front` parameter to
distinguish between applying the op to the front of the circuit. It also takes
an `out_node` pointer parameter which is written with the value of the node
added to the graph on success. All iterators will yield indices from the DAG a
well (doesn't exist yet).

Generally, the approach to accessing node weights in the DAG is to call one of
the accessor functions with a pointer to the DAG and the index of the node you
want to access. Functions like `qk_dag_apply_gate` return the index of the node
that was added for use in future access.

The `qk_dag_op_node_kind` function returns a new QkOperationKind enum
that gives callers the operation's kind for use in dispatching
operation-specific calls, e.g. `qk_dag_op_node_gate_op` to retrieve the
gate and params of a standard gate (QkOperationKind_Gate).

Our Rust `Qubit` and `Clbit` types are now also `#[repr(transparent)]` so that
we can reinterpret them as `u32`s, and thus return pointers to the buffers
owned by our interners for bits, avoiding a copy.