Split VF2 `call_limit` to before and after first layout (#14862)
* Split VF2 `call_limit` to before and after first layout

With the new VF2 scoring algorithm, the `max_trials` option is no longer
appropriate as a limit, but is one that we functionally relied on for
good runtimes in the face of high numbers of isomorphisms.  This splits
the `call_limit` functionality in two, so that a separate (likely lower)
limit can be placed on the time spent improving the layout, while
retaining a high call limit for the initial search.

* Fix call-limit unbound conversion

* Remove unnecessary pub

* Revert separation of O1 and O2 VF2 configuration

That shouldn't be being done in this PR, if at all.

* Clarify `call_limit` meaning

Previously it was not clear that it's just two limits and we swap from
one to the other at the point of the first match.  The previous switch
was also not entirely correct, since swapping from a `None` to a `Some`
limit would not have counted calls already made.